+++ src/demos/cards/card_ui.cljs
@@ -3,8 +3,9 @@
     devcards.core
     cards.A-Introduction
     cards.autocomplete-cards
-    cards.component-local-state-cards
     cards.background-load-cards
+    cards.component-local-state-cards
+    cards.dynamic-routing-cards
     cards.error-handling-cards
     cards.initial-state-cards
     cards.lazy-loading-indicators-cards